+++ libs/btp/tool-layout/src/directives/tool-layout-container.directive.ts
@@ -1,0 +1,10 @@
+import { Directive } from '@angular/core';
+
+@Directive({
+    selector: '[fdbToolLayoutContainer]',
+    standalone: true,
+    host: {
+        class: 'fd-tool-layout__container'
+    }
+})
+export class ToolLayoutContainerDirective {}