+++ libs/docs/btp/tool-layout/tool-layout-header/tool-layout-header.component.ts
@@ -1,0 +1,25 @@
+import { Component } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
+import {
+    DescriptionComponent,
+    HeaderComponent,
+    HeaderTabsComponent,
+    ImportComponent
+} from '@fundamental-ngx/docs/shared';
+
+@Component({
+    templateUrl: './tool-layout-header.component.html',
+    styles: [
+        `
+            code {
+                color: #00677f;
+            }
+        `
+    ],
+
+    standalone: true,
+    imports: [HeaderComponent, DescriptionComponent, ImportComponent, HeaderTabsComponent, RouterOutlet]
+})
+export class ToolLayoutHeaderComponent {
+    constructor() {}
+}